mythrilmerch-backend/requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
Flask-Limiter==3.5.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.10
psycopg2-pool==1.1
python-dotenv==1.0.0
gunicorn==21.2.0
prometheus-client==0.19.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
requests==2.31.0 